Process primitive extensions on Questionnaire root for FHIR↔FCE.
Extract shared helpers so fields like _title convert the same way as item-level _text.
